+++ CHANGES.md
@@ -20,6 +20,7 @@ Changes in 10.9
 - CSS Add modname class, ie .modname-grades-reportcards-php for modname=Grades/ReportCards.php in PDF.php
 - CSS Add .list-column-[column_name] class in ListOutput.fnc.php
 - CSS Set Report Cards Comments column (max) width to 36% in wkhtmltopdf.css
+- Temporary AllowEdit so SelectInput() is displayed to everyone in DailyTransactions.php
 
 Changes in 10.8.4
 -----------------